답변 목록 조회 개발

+++ src/main/java/kr/co/puerpuella/apitextssul/model/entity/ArticleComment.java
@@ -21,7 +21,7 @@ public class ArticleComment extends LabelEntity{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "COMMENT_ID")
-    private Long commentId;
+    private Integer commentId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "ARTICLE_ID")
